changed virtuoso_perception to cpp to add euclidean clustering
