+++ virtuoso_msgs/msg/BoundingBox.msg
@@ -1,0 +1,25 @@
+# Essentially a copy of Autoware.auto's BoundingBox message
+
+geometry_msgs/Point32 centroid
+
+geometry_msgs/Point32 size
+
+geometry_msgs/Quaternion orientation
+
+float32 velocity
+
+float32 heading
+
+float32 heading_rate
+
+geometry_msgs/Point32[] corners
+
+float32 variance
+
+float32 value
+
+uint8 vehicle_label
+
+uint8 signal_label
+
+float32 class_likelihood
