+++ virtuoso_msgs/msg/BoundingBoxArray.msg
@@ -1,0 +1,5 @@
+# Essentially a copy of Autoware.auto's BoundingBoxArray
+
+std_msgs/Header header
+
+virtuoso_msgs/BoundingBox[] boxes
