+++ virtuoso_perception/launch/find_and_classify_buoys.launch.py
@@ -9,11 +9,10 @@ def generate_launch_description():
 
     pkg_share = get_package_share_directory('virtuoso_perception')
 
-    # must have virtuoso_processing lidar_processing launched as well for euclidean clustering to do anything
     return LaunchDescription([
         IncludeLaunchDescription(PythonLaunchDescriptionSource(os.path.join(pkg_share, 'launch/find_buoys.launch.py'))),
         Node(
             package='virtuoso_perception',
-            executable='classify_buoys'
+            executable='classify_buoys_node.py'
         )        
     ])
