+++ virtuoso_perception/scripts/lidar_processing/self_filter_node.py
@@ -1,8 +1,10 @@
+#!/usr/bin/env python3
+
 import rclpy
 from rclpy.node import Node
 from sensor_msgs.msg import PointCloud2
 import math
-from ..utils.pointcloud import read_points, create_cloud_xyz32
+from virtuoso_perception.utils.pointcloud import read_points, create_cloud_xyz32
 
 class SelfFilter(Node):
 
@@ -13,7 +15,7 @@ class SelfFilter(Node):
             '/perception/lidar/points_self_filtered', 10)
 
         self.declare_parameters(namespace='', parameters=[
-            ('radius', 0)
+            ('radius', 0.0)
         ])
 
     def callback(self, msg:PointCloud2):
