+++ virtuoso_perception/scripts/stereo/buoy_stereo_node.py
@@ -1,14 +1,16 @@
+#!/usr/bin/env python3
+
 import rclpy
 from rclpy.node import Node
-from .buoy_stereo import BuoyStereo
-from virtuoso_msgs.msg import Contours, BuoyArray
+from virtuoso_perception.stereo.buoy_stereo import BuoyStereo
+from virtuoso_msgs.msg import BuoyArray
+from virtuoso_msgs.srv import ImageBuoyStereo, ImageBuoyFilter
 from sensor_msgs.msg import CameraInfo, PointCloud2, Image
 from geometry_msgs.msg import TransformStamped
 from tf2_ros.buffer import Buffer
 from tf2_ros.transform_listener import TransformListener
 from rclpy.time import Time
-from .utils import tf_transform_to_cv2_transform
-from virtuoso_msgs.srv import ImageBuoyStereo, ImageBuoyFilter
+from virtuoso_perception.stereo.utils import tf_transform_to_cv2_transform
 from rclpy.executors import MultiThreadedExecutor
 from rclpy.callback_groups import MutuallyExclusiveCallbackGroup
 import time
