+++ virtuoso_perception/scripts/testing/buoy_cam_filter.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+
 import rclpy
 from rclpy.node import Node
 from sensor_msgs.msg import Image, CameraInfo
