+++ virtuoso_perception/scripts/testing/find_dock_codes.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+
 import rclpy
 from rclpy.node import Node
 from virtuoso_msgs.srv import DockCodesCameraPos
