--- virtuoso_perception/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/virtuoso_perception
-[install]
-install-scripts=$base/lib/virtuoso_perception
