+++ virtuoso_perception/virtuoso_perception/dock/find_dock_posts.py
@@ -1,13 +1,12 @@
-from ..utils.color_range import ColorRange
-from ..utils.ColorFilter import ColorFilter
-from ..clustering.density_filter import DensityFilter
-from ..utils.node_helper import NodeHelper
-from ..stereo.utils import unflatten_contours
+from virtuoso_perception.utils.color_range import ColorRange
+from virtuoso_perception.utils.ColorFilter import ColorFilter
+from virtuoso_perception.clustering.density_filter import DensityFilter
+from virtuoso_perception.utils.node_helper import NodeHelper
+from virtuoso_perception.stereo.utils import unflatten_contours
 import numpy as np
 from sensor_msgs.msg import Image
 from cv_bridge import CvBridge
 import cv2
-import random
 
 class FindDockPosts(NodeHelper):
