+++ virtuoso_perception/virtuoso_perception/stereo/stereo.py
@@ -1,11 +1,9 @@
-from sensor_msgs.msg import CameraInfo, PointCloud2
-from virtuoso_msgs.msg import Contours, BuoyArray
+from sensor_msgs.msg import CameraInfo
 from .utils import contour_average_yx
 import numpy as np
 from rclpy.node import Node
 import cv2
-from ..utils.node_helper import NodeHelper
-from virtuoso_perception.utils.pointcloud import create_cloud_xyz32
+from virtuoso_perception.utils.node_helper import NodeHelper
 from cv_bridge import CvBridge
 
 class Stereo(NodeHelper):
